--- src/main/java/myplace/core/member/MemberRepository.java
@@ -1,9 +1,0 @@
-package myplace.core.member;
-
-import java.util.Optional;
-
-public interface MemberRepository {
-    Member save(Member member);
-    Optional<Member> findById(Long id);
-    Optional<Member> findByUsername(String username);
-}
